Add system function and configuration parameter
Added a new function class "SystemFunction" in functions.py to execute a command on the local host and return the output. This provides additional functionality to the application in case it's needed to execute system commands. As it's a sensitive feature, it's controlled by a flag/configuration parameter "system_function_enabled" added in cfg_example.json. Also, made appropriate changes in main.py to add the SystemFunction to available functions only if it's enabled via config. Lastly, added method descriptor() in 'DatetimeFunction' and 'SystemFunction' for better code structure.